stop leaving temp files in CWD with MSVC
A temp, a .def, a .pdb and a .obj were left in the CWD and never deleted
previously. By supply explict paths to VC these files will be placed in
the temp dir that HasCompiler created and deleted at the same time as all
other temp files that HasCompiler creates.